+++ flutter_modular/lib/flutter_modular.dart
@@ -2,9 +2,9 @@ library flutter_modular;
 
 export 'package:flutter_modular_annotations/flutter_modular_annotations.dart';
 
-export 'src/core/interfaces/child_module.dart';
 export 'src/core/interfaces/disposable.dart';
 export 'src/core/interfaces/modular_route.dart';
+export 'src/core/interfaces/module.dart';
 export 'src/core/interfaces/route_guard.dart';
 export 'src/core/models/bind.dart';
 export 'src/core/models/child_route.dart';
@@ -13,7 +13,6 @@ export 'src/core/models/modular_arguments.dart';
 export 'src/core/models/module_route.dart';
 export 'src/core/models/wildcard_route.dart';
 export 'src/presenters/inject.dart';
-export 'src/presenters/main_module.dart';
 export 'src/presenters/modular_base.dart';
 export 'src/presenters/widgets/modular_app.dart';
 export 'src/presenters/widgets/modular_state.dart';