+++ flutter_modular/lib/src/core/interfaces/modular_interface.dart
@@ -1,14 +1,14 @@
 import '../models/modular_arguments.dart';
-import 'child_module.dart';
+import 'module.dart';
 import 'modular_navigator_interface.dart';
 
 abstract class ModularInterface {
   bool get debugMode;
   ModularArguments? get args;
   String get initialRoute;
-  ChildModule get initialModule;
-  void init(ChildModule module);
-  void bindModule(ChildModule module, [String path]);
+  Module get initialModule;
+  void init(Module module);
+  void bindModule(Module module, [String path]);
   void debugPrintModular(String text);
 
   IModularNavigator get to;