+++ flutter_modular/lib/src/core/interfaces/modular_route.dart
@@ -5,16 +5,14 @@ import 'package:flutter/widgets.dart';
 
 import '../models/custom_transition.dart';
 import '../models/modular_arguments.dart';
-import 'child_module.dart';
+import 'module.dart';
 import 'route_guard.dart';
 
-typedef RouteBuilder<T> = Route<T> Function(
-    WidgetBuilder, RouteSettings);
-typedef ModularChild = Widget Function(
-    BuildContext context, ModularArguments? args);
+typedef RouteBuilder<T> = Route<T> Function(WidgetBuilder, RouteSettings);
+typedef ModularChild = Widget Function(BuildContext context, ModularArguments? args);
 
 abstract class ModularRoute<T> {
-  ChildModule? get currentModule;
+  Module? get currentModule;
 
   ModularArguments? get args;
 
@@ -54,11 +52,11 @@ abstract class ModularRoute<T> {
   ///
   /// The module will be loaded
   ///
-  /// Type: ChildModule
+  /// Type: Module
   ///
   /// For more example go to Modular page from gitHub [https://github.com/Flutterando/modular]
   ///
-  ChildModule? get module;
+  Module? get module;
 
   ///
   /// Paramenter name: [params]
@@ -90,7 +88,7 @@ abstract class ModularRoute<T> {
   /// Type: Map<String, String>
   ///
   /// For more example http://example.com/help?id=12&rate=22
-  /// 
+  ///
   Map<String, List<String>> get queryParamsAll;
 
   ///
@@ -204,38 +202,20 @@ abstract class ModularRoute<T> {
   Map<
       TransitionType,
       PageRouteBuilder<T> Function(
-          Widget Function(BuildContext, ModularArguments?) builder,
-          ModularArguments? args,
-          Duration transitionDuration,
-          RouteSettings settings,
-          )> get transitions;
-
-  ModularRoute<T> copyWith({ModularChild? child,
-    String? routerName,
-    ChildModule? module,
-    List<ModularRoute>? children,
-    List<ModularRoute>? routerOutlet,
-    ChildModule? currentModule,
-    Map<String, String>? params,
-    List<RouteGuard>? guards,
-    TransitionType? transition,
-    RouteBuilder<T>? routeGenerator,
-    String? modulePath,
-    Uri uri,
-    Duration? duration,
-    Completer<T>? popRoute,
+    Widget Function(BuildContext, ModularArguments?) builder,
     ModularArguments? args,
-    CustomTransition? customTransition});
+    Duration transitionDuration,
+    RouteSettings settings,
+  )> get transitions;
+
+  ModularRoute<T> copyWith({ModularChild? child, String? routerName, Module? module, List<ModularRoute>? children, List<ModularRoute>? routerOutlet, Module? currentModule, Map<String, String>? params, List<RouteGuard>? guards, TransitionType? transition, RouteBuilder<T>? routeGenerator, String? modulePath, Uri uri, Duration? duration, Completer<T>? popRoute, ModularArguments? args, CustomTransition? customTransition});
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
   bool operator ==(Object o) {
     if (identical(this, o)) return true;
 
-    return o is ModularRoute<T> &&
-        o.modulePath == modulePath &&
-        o.routerName == routerName &&
-        o.module == module;
+    return o is ModularRoute<T> && o.modulePath == modulePath && o.routerName == routerName && o.module == module;
   }
 
   @override